fsd project/ml_service/requirements.txt
# ML Service Requirements
# Python 3.9+

# Core ML Libraries (Baseline Model)
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
shap>=0.43.0

# API Framework
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Data Processing
joblib>=1.3.0

# Optional: For advanced models (uncomment when upgrading)
# xgboost>=2.0.0
# lightgbm>=4.0.0

# Optional: Utilities
# requests>=2.31.0
# gunicorn>=21.0.0
# python-json-logger>=2.0.0

# Optional: Testing
# pytest>=7.4.0
# pytest-cov>=4.1.0
